requirements.txt
asgiref==3.8.1
Django==5.1.1
django-ckeditor==6.7.1
django-cuser==2017.3.16
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.17
django-phonenumber-field==8.0.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
modeltranslation==0.25
packaging==24.1
phonenumbers==8.13.46
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
